// LibeyTechnicalTestWeb/src/app/models/user.ts
export class User {
    documentNumber: string ='';
    documentTypeId: number =0;
    name: string='';
    fathersLastName: string='';
    mothersLastName: string='';
    address: string='';
    ubigeoCode: string='';
    phone: string='';
    email: string='';
    password: string='';
    active: boolean=true;
  }